+++ extensions/pins/src/index.tsx
@@ -1,7 +1,22 @@
 import { useEffect, useState } from "react";
-import { MenuBarExtra, open, openExtensionPreferences, getPreferenceValues, getFrontmostApplication } from "@raycast/api";
+import {
+  MenuBarExtra,
+  open,
+  openExtensionPreferences,
+  getPreferenceValues,
+  getFrontmostApplication,
+} from "@raycast/api";
 import { getFavicon, useCachedState } from "@raycast/utils";
-import { setStorage, getStorage, iconMap, useGroups, copyPinData, openPin, createNewPin, getCurrentDirectory } from "./utils";
+import {
+  setStorage,
+  getStorage,
+  iconMap,
+  useGroups,
+  copyPinData,
+  openPin,
+  createNewPin,
+  getCurrentDirectory,
+} from "./utils";
 import { StorageKey } from "./constants";
 import { Pin, Group, ExtensionPreferences } from "./types";
 import { SupportedBrowsers, getCurrentURL } from "./browser-utils";
@@ -48,7 +63,7 @@ const getGroupIcon = (groupName: string, groups: Group[]) => {
 export default function Command() {
   const { pinGroups, isLoading } = usePinGroups();
   const [groups, setGroups] = useGroups();
-  const [currentAppName, setCurrentAppName] = useCachedState<string>("current-app-name","");
+  const [currentAppName, setCurrentAppName] = useCachedState<string>("current-app-name", "");
   const preferences = getPreferenceValues<ExtensionPreferences>();
   const pinIcon = { source: { light: "pin-icon.svg", dark: "pin-icon@dark.svg" } };
 
@@ -106,7 +121,7 @@ export default function Command() {
           </MenuBarExtra.Section>
         ) : null}
 
-        {preferences.showCategories && groups?.length ? (
+        {preferences.showCategories && groups?.length && Object.keys(usedGroups).length ? (
           <MenuBarExtra.Section title="Groups">
             {Object.keys(usedGroups).map((key) => (
               <MenuBarExtra.Submenu title={key} key={key} icon={getGroupIcon(key, groups as Group[])}>
@@ -127,32 +142,48 @@ export default function Command() {
                   />
                 ))}
 
-                <MenuBarExtra.Separator />
-                {preferences.showOpenAll ? (
-                  <MenuBarExtra.Item
-                    title="Open All"
-                    onAction={() => usedGroups[key].forEach((pin: Pin) => open(pin.url))}
-                  />
-                ) : null}
+                <MenuBarExtra.Section>
+                  {preferences.showOpenAll ? (
+                    <MenuBarExtra.Item
+                      title="Open All"
+                      onAction={() => usedGroups[key].forEach((pin: Pin) => openPin(pin, preferences))}
+                    />
+                  ) : null}
+                </MenuBarExtra.Section>
               </MenuBarExtra.Submenu>
             ))}
           </MenuBarExtra.Section>
         ) : null}
 
         <MenuBarExtra.Section>
-          {preferences.showPinShortcut && SupportedBrowsers.includes(currentAppName) ? <MenuBarExtra.Item title="Pin This Tab" onAction={async () => {
-            const currentTab = await getCurrentURL(currentAppName);
-            const tabName = currentTab[0]
-            const tabURL = currentTab[1]
-            await createNewPin(tabName, tabURL, "Favicon / File Icon", "None")
-          }} /> : null}
-          {preferences.showPinShortcut && currentAppName == "Finder" ? <MenuBarExtra.Item title={`Pin This Directory`} tooltip="Add a pin whose target path is the current directory of Finder" onAction={async () => {
-            const currentDir = await getCurrentDirectory();
-            const dirName = currentDir[0]
-            const dirPath = currentDir[1]
-            await createNewPin(dirName, dirPath, "Favicon / File Icon", "None")
-          }} /> : null}
-          <MenuBarExtra.Item title="Copy Pin Data" tooltip="Copy the JSON data for all of your pins" onAction={() => copyPinData()} />
+          {preferences.showPinShortcut && SupportedBrowsers.includes(currentAppName) ? (
+            <MenuBarExtra.Item
+              title="Pin This Tab"
+              onAction={async () => {
+                const currentTab = await getCurrentURL(currentAppName);
+                const tabName = currentTab[0];
+                const tabURL = currentTab[1];
+                await createNewPin(tabName, tabURL, "Favicon / File Icon", "None");
+              }}
+            />
+          ) : null}
+          {preferences.showPinShortcut && currentAppName == "Finder" ? (
+            <MenuBarExtra.Item
+              title={`Pin This Directory`}
+              tooltip="Add a pin whose target path is the current directory of Finder"
+              onAction={async () => {
+                const currentDir = await getCurrentDirectory();
+                const dirName = currentDir[0];
+                const dirPath = currentDir[1];
+                await createNewPin(dirName, dirPath, "Favicon / File Icon", "None");
+              }}
+            />
+          ) : null}
+          <MenuBarExtra.Item
+            title="Copy Pin Data"
+            tooltip="Copy the JSON data for all of your pins"
+            onAction={() => copyPinData()}
+          />
           <MenuBarExtra.Item title="Preferences..." onAction={() => openExtensionPreferences()} />
         </MenuBarExtra.Section>
       </MenuBarExtra>