+++ extensions/pins/src/view-pins.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from "react";
 import {
   Icon,
-  open,
   Form,
   List,
   useNavigation,
@@ -42,11 +41,12 @@ const modifyPin = async (
   icon: string,
   group: string,
   pop: () => void,
-  setPins: (pins: Pin[]) => void
+  setPins: React.Dispatch<React.SetStateAction<Pin[] | undefined>>
 ) => {
   const storedPins = await getStorage(StorageKey.LOCAL_PINS);
 
-  const newData = storedPins.map((oldPin: Pin) => {
+  const newData: Pin[] = storedPins.map((oldPin: Pin) => {
+    // Update pin if it exists
     if (oldPin.id == pin.id) {
       return {
         name: name,
@@ -60,13 +60,27 @@ const modifyPin = async (
     }
   });
 
+  if (newData.every((storedPin) => storedPin.id != pin.id)) {
+    // Add new pin if it doesn't exist
+    newData.push({
+      name: name,
+      url: url,
+      icon: icon,
+      group: group,
+      id: pin.id,
+    });
+  }
+
   setPins(newData);
   await setStorage(StorageKey.LOCAL_PINS, newData);
   await showToast({ title: `Updated pin!` });
   pop();
 };
 
-const EditPinView = (props: { pin: Pin; setPins: (pins: Pin[]) => void }) => {
+const EditPinView = (props: {
+  pin: Pin;
+  setPins: React.Dispatch<React.SetStateAction<Pin[] | undefined>>;
+}) => {
   const pin = props.pin;
   const setPins = props.setPins;
   const [urlError, setUrlError] = useState<string | undefined>();
@@ -105,7 +119,7 @@ const EditPinView = (props: { pin: Pin; setPins: (pins: Pin[]) => void }) => {
         onBlur={(event) => {
           if (event.target.value?.length == 0) {
             setUrlError("URL cannot be empty!");
-          } else if (!event.target.value?.includes(":") && !event.target.value?.startsWith("/")) {
+          } else if (!event.target.value?.includes(":") && !event.target.value?.startsWith("/") && !event.target.value?.startsWith("~")) {
             setUrlError("Please enter a valid URL or path!");
           } else if (urlError !== undefined) {
             setUrlError(undefined);
@@ -140,7 +154,7 @@ const EditPinView = (props: { pin: Pin; setPins: (pins: Pin[]) => void }) => {
   );
 };
 
-const deletePin = async (pin: Pin, setPins: (pins: Pin[]) => void) => {
+const deletePin = async (pin: Pin, setPins: React.Dispatch<React.SetStateAction<Pin[] | undefined>>) => {
   if (await confirmAlert({ title: "Are you sure?" })) {
     const storedPins = await getStorage(StorageKey.LOCAL_PINS);
 
@@ -154,7 +168,7 @@ const deletePin = async (pin: Pin, setPins: (pins: Pin[]) => void) => {
   }
 };
 
-const movePinUp = async (index: number, setPins: (pins: Pin[]) => void) => {
+const movePinUp = async (index: number, setPins: React.Dispatch<React.SetStateAction<Pin[] | undefined>>) => {
   const storedPins = await getStorage(StorageKey.LOCAL_PINS);
   if (storedPins.length > index) {
     [storedPins[index - 1], storedPins[index]] = [storedPins[index], storedPins[index - 1]];
@@ -163,7 +177,7 @@ const movePinUp = async (index: number, setPins: (pins: Pin[]) => void) => {
   }
 };
 
-const movePinDown = async (index: number, setPins: (pins: Pin[]) => void) => {
+const movePinDown = async (index: number, setPins: React.Dispatch<React.SetStateAction<Pin[] | undefined>>) => {
   const storedPins = await getStorage(StorageKey.LOCAL_PINS);
   if (storedPins.length > index + 1) {
     [storedPins[index], storedPins[index + 1]] = [storedPins[index + 1], storedPins[index]];
@@ -172,13 +186,41 @@ const movePinDown = async (index: number, setPins: (pins: Pin[]) => void) => {
   }
 };
 
+const CreateNewPinAction = (props: {
+  setPins: React.Dispatch<React.SetStateAction<Pin[] | undefined>>;
+}) => {
+  const { setPins } = props;
+  return (
+    <Action.Push
+      title="Create New Pin"
+      icon={Icon.PlusCircle}
+      shortcut={{ modifiers: ["cmd"], key: "n" }}
+      target={
+        <EditPinView
+          pin={{
+            name: "",
+            url: "",
+            icon: "None",
+            group: "None",
+            id: -1,
+          }}
+          setPins={setPins}
+        />
+      }
+    />
+  );
+}
+
 export default function Command() {
-  const [pins, setPins] = usePins();
-  const { push } = useNavigation();
+  const { pins, setPins } = usePins();
   const preferences = getPreferenceValues<ExtensionPreferences>();
 
   return (
-    <List isLoading={pins === undefined} searchBarPlaceholder="Search pins...">
+    <List
+      isLoading={pins === undefined}
+      searchBarPlaceholder="Search pins..."
+      actions={<ActionPanel>{<CreateNewPinAction setPins={setPins} />}</ActionPanel>}
+    >
       <List.EmptyView title="No Pins Found" icon="no-view.png" />
       {pins &&
         (pins as Pin[]).map((pin, index) => (
@@ -197,43 +239,56 @@ export default function Command() {
             }
             actions={
               <ActionPanel>
-                <Action title="Open" icon={Icon.ChevronRight} onAction={() => openPin(pin, preferences)} />
-                <Action
-                  title="Edit"
-                  icon={Icon.Pencil}
-                  onAction={() => push(<EditPinView pin={pin} setPins={setPins as (pins: Pin[]) => void} />)}
-                />
+                <ActionPanel.Section title="Pin Actions">
+                  <Action title="Open" icon={Icon.ChevronRight} onAction={() => openPin(pin, preferences)} />
+                  <Action.Push
+                    title="Edit"
+                    icon={Icon.Pencil}
+                    shortcut={{ modifiers: ["cmd"], key: "e" }}
+                    target={<EditPinView pin={pin} setPins={setPins} />}
+                  />
+                  <Action.Push
+                    title="Duplicate"
+                    icon={Icon.EyeDropper}
+                    shortcut={{ modifiers: ["cmd", "ctrl"], key: "d" }}
+                    target={<EditPinView pin={{...pin, name: pin.name + " Copy"}} setPins={setPins} />}
+                  />
 
-                {index > 0 ? (
+                  {index > 0 ? (
+                    <Action
+                      title="Move Up"
+                      icon={Icon.ArrowUp}
+                      shortcut={{ modifiers: ["cmd", "shift"], key: "u" }}
+                      onAction={() => {
+                        movePinUp(index, setPins);
+                        clearSearchBar();
+                      }}
+                    />
+                  ) : null}
+                  {index < pins.length - 1 ? (
+                    <Action
+                      title="Move Down"
+                      icon={Icon.ArrowDown}
+                      shortcut={{ modifiers: ["cmd", "shift"], key: "d" }}
+                      onAction={() => {
+                        movePinDown(index, setPins);
+                        clearSearchBar();
+                      }}
+                    />
+                  ) : null}
+
                   <Action
-                    title="Move Up"
-                    icon={Icon.ArrowUp}
+                    title="Delete Pin"
+                    icon={Icon.Trash}
                     onAction={() => {
-                      movePinUp(index, setPins as (pins: Pin[]) => void);
+                      deletePin(pin, setPins);
                       clearSearchBar();
                     }}
+                    shortcut={{ modifiers: ["cmd"], key: "d" }}
+                    style={Action.Style.Destructive}
                   />
-                ) : null}
-                {index < pins.length - 1 ? (
-                  <Action
-                    title="Move Down"
-                    icon={Icon.ArrowDown}
-                    onAction={() => {
-                      movePinDown(index, setPins as (pins: Pin[]) => void);
-                      clearSearchBar();
-                    }}
-                  />
-                ) : null}
-
-                <Action
-                  title="Delete Pin"
-                  icon={Icon.Trash}
-                  onAction={() => {
-                    deletePin(pin, setPins as (pins: Pin[]) => void);
-                    clearSearchBar();
-                  }}
-                  style={Action.Style.Destructive}
-                />
+                </ActionPanel.Section>
+                <CreateNewPinAction setPins={setPins} />
               </ActionPanel>
             }
           />